store: Protect from deletion of local key via kvstore event
Ensure that local keys can never be removed via a kvstore delete event.  The
most likely reason for this to happen is if the lease of an old agent expires
before the new agent with a matching local key manages to update the key and
establish a new lease.

Fixes: #7222

+++ pkg/kvstore/store/store.go
@@ -293,6 +293,19 @@ func (s *SharedStore) syncLocalKeys() error {
 	return nil
 }
 
+func (s *SharedStore) lookupLocalKey(name string) LocalKey {
+	s.mutex.RLock()
+	defer s.mutex.RUnlock()
+
+	for _, key := range s.localKeys {
+		if key.GetKeyName() == name {
+			return key
+		}
+	}
+
+	return nil
+}
+
 // UpdateLocalKey adds a key to be synchronized with the kvstore
 func (s *SharedStore) UpdateLocalKey(key LocalKey) {
 	s.mutex.Lock()
@@ -316,13 +329,14 @@ func (s *SharedStore) UpdateLocalKeySync(key LocalKey) error {
 
 // DeleteLocalKey removes a key from being synchronized with the kvstore
 func (s *SharedStore) DeleteLocalKey(key LocalKey) {
-	err := s.backend.Delete(s.keyPath(key))
 	name := key.GetKeyName()
 
 	s.mutex.Lock()
 	_, ok := s.localKeys[name]
 	delete(s.localKeys, name)
 	s.mutex.Unlock()
+
+	err := s.backend.Delete(s.keyPath(key))
 
 	if ok {
 		if err != nil {
@@ -446,7 +460,12 @@ func (s *SharedStore) watcher(listDone chan bool) {
 				}
 
 			case kvstore.EventTypeDelete:
-				s.deleteKey(keyName)
+				if localKey := s.lookupLocalKey(keyName); localKey != nil {
+					logger.Warning("Received delete event for local key. Re-creating the key in the kvstore")
+					s.syncLocalKey(localKey)
+				} else {
+					s.deleteKey(keyName)
+				}
 			}
 		}
 	}
